frontend/Rythu-Mitra-main(1)/Rythu-Mitra-main/backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
gunicorn>=20.1.0
psycopg2-binary>=2.9.0
celery>=5.3.0
redis>=4.5.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
xgboost>=1.7.0
lightgbm>=3.3.5
statsmodels>=0.14.0
prophet>=1.1.0
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.12.0
Pillow>=9.5.0
requests>=2.31.0
boto3>=1.26.0
sentry-sdk>=1.25.0
python-dotenv>=1.0.0
drf-yasg>=1.21.0
whitenoise>=6.5.0
dj-database-url>=2.1.0
